move cmdtest & wellknown
