+++ cmd/jimmctl/cmd/grantauditlogaccess_test.go
@@ -7,7 +7,7 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/canonical/jimm/v3/cmd/jimmctl/cmd"
-	"github.com/canonical/jimm/v3/internal/cmdtest"
+	"github.com/canonical/jimm/v3/internal/testutils/cmdtest"
 )
 
 type grantAuditLogAccessSuite struct {
